feat: exceptionPayload customization for fields validation
