+++ src/main/java/br/com/meli/dhprojetointegrador/exception/handler/ControllerAdvisor.java
@@ -1,36 +1,78 @@
 package br.com.meli.dhprojetointegrador.exception.handler;
 
-import java.time.LocalDateTime;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
+import br.com.meli.dhprojetointegrador.dto.response.ExceptionPayloadDTO;
 import br.com.meli.dhprojetointegrador.dto.response.ExceptionPayloadResponse;
 import br.com.meli.dhprojetointegrador.exception.BusinessValidatorException;
+import br.com.meli.dhprojetointegrador.exception.PurchaseOrderNotFoundException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.time.LocalDateTime;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
- *  Used to catch and return the proper response payload for exceptions thrown at the layers bellow the controllers.
+ * Used to catch and return the proper response payload for exceptions thrown at the layers bellow the controllers.
  */
 @RestControllerAdvice
 public class ControllerAdvisor extends ResponseEntityExceptionHandler {
 
+    /**
+     * @param exception BusinessValidator thrown at the service layer
+     * @return ExceptionPayloadResponse filled with the exception data
+     */
+    @ExceptionHandler(BusinessValidatorException.class)
+    protected ResponseEntity<?> handleAuthException(BusinessValidatorException exception) {
+        ExceptionPayloadResponse exceptionPayload = ExceptionPayloadResponse.builder()
+                .timestamp(LocalDateTime.now())
+                .title(exception.getErrorTitle())
+                .statusCode(exception.getHttpStatus().value())
+                .description(exception.getMessage())
+                .build();
+
+        return new ResponseEntity<>(exceptionPayload, exception.getHttpStatus());
+    }
+
 	/**
-	 *
-	 * @param exception BusinessValidator thrown at the service layer
-	 * @return ExceptionPayloadResponse filled with the exception data
+	 * Generates payload response for exceptions related to field validation on request bodies
 	 */
-	@ExceptionHandler(BusinessValidatorException.class)
-	    protected ResponseEntity<?> handleAuthException(BusinessValidatorException exception) {
-			ExceptionPayloadResponse exceptionPayload = ExceptionPayloadResponse.builder()
-					.timestamp(LocalDateTime.now())
-					.title(exception.getErrorTitle())
-					.statusCode(exception.getHttpStatus().value())
-					.description(exception.getMessage())
-					.build();
+	@Override
+	protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
+		BindingResult bindingResult = ex.getBindingResult();
 
-			return new ResponseEntity<>(exceptionPayload, exception.getHttpStatus());
-		}
+		Map<String, String> errorsMap = bindingResult
+				.getFieldErrors()
+				.stream()
+				.collect(Collectors.toMap(FieldError::getField, FieldError::getDefaultMessage));
 
+		ExceptionPayloadResponse exceptionPayload = ExceptionPayloadResponse.builder()
+				.timestamp(LocalDateTime.now())
+				.title("Some fields are missing in the request payload")
+				.statusCode(HttpStatus.BAD_REQUEST.value())
+				.fieldToMessageMap(errorsMap)
+				.description(String.format("A total of %d fields are missing", bindingResult.getErrorCount()))
+				.build();
+
+		return new ResponseEntity<>(exceptionPayload, HttpStatus.UNPROCESSABLE_ENTITY);
+	}
+
+	@ExceptionHandler(value = {PurchaseOrderNotFoundException.class})
+	protected ResponseEntity<Object> handlePurchaseOrderNotFoundException(PurchaseOrderNotFoundException exception) {
+		ExceptionPayloadDTO exceptionPayload = ExceptionPayloadDTO.builder()
+				.timestamp(LocalDateTime.now())
+				.title("PurchaseOrder Not Found")
+				.statusCode(HttpStatus.NOT_FOUND.value())
+				.description(exception.getMessage())
+				.build();
+
+		return new ResponseEntity<>(exceptionPayload, HttpStatus.NOT_FOUND);
+	}
 }
